+++ base/integration_test.go
@@ -1,0 +1,9 @@
+package base
+
+import (
+	"testing"
+)
+
+func TestMain(t *testing.M) {
+	NewDockerTest(t)
+}
