+++ rest/integration_test.go
@@ -1,0 +1,11 @@
+package rest
+
+import (
+	"testing"
+
+	"github.com/couchbase/sync_gateway/base"
+)
+
+func TestMain(t *testing.M) {
+	base.NewDockerTest(t)
+}
