+++ test-integration-init.sh
@@ -15,7 +15,7 @@ CB_ADMIN_USERNAME="Administrator"
 CB_ADMIN_PASSWORD="password"
 
 SG_TEST_BUCKETS=("test_data_bucket" "test_shadowbucket" "test_indexbucket")
-SG_TEST_BUCKET_RAMSIZE=1000 # MB
+SG_TEST_BUCKET_RAMSIZE=768 # MB
 
 SG_TEST_BUCKET_PASSWORD="password"
 SG_TEST_BUCKET_RBAC_ROLES=() # No bucket-specific roles when we can rely on global admin
